fix(variables): move `getAdHocFiltersVariable`
