+++ src/services/scenes.ts
@@ -1,12 +1,6 @@
 import { urlUtil } from '@grafana/data';
 import { DataSourceWithBackend, config, getDataSourceSrv } from '@grafana/runtime';
-import {
-  AdHocFiltersVariable,
-  getUrlSyncManager,
-  sceneGraph,
-  SceneObject,
-  SceneObjectUrlValues,
-} from '@grafana/scenes';
+import { getUrlSyncManager, sceneGraph, SceneObject, SceneObjectUrlValues } from '@grafana/scenes';
 import { VAR_DATASOURCE_EXPR, LOG_STREAM_SELECTOR_EXPR } from './variables';
 import { EXPLORATIONS_ROUTE } from './routing';
 import { IndexScene } from 'Components/IndexScene/IndexScene';
@@ -41,22 +35,6 @@ export async function getLokiDatasource(sceneObject: SceneObject) {
     | DataSourceWithBackend
     | undefined;
   return ds;
-}
-
-export function getAdHocFiltersVariable(variableName: string, sceneObject: SceneObject) {
-  const variable = sceneGraph.lookupVariable(variableName, sceneObject);
-
-  if (!variable) {
-    console.warn(`Could not get AdHocFiltersVariable ${variableName}. Variable not found.`);
-    return null;
-  }
-  if (!(variable instanceof AdHocFiltersVariable)) {
-    console.warn(
-      `Could not get AdHocFiltersVariable ${variableName}. Variable is not an instance of AdHocFiltersVariable`
-    );
-    return null;
-  }
-  return variable;
 }
 
 export function isDefined<T>(value: T | null | undefined): value is T {
